refactor : PerformanceDate 조회 로직을 PerformanceRepository에서 PerformanceDateRepository로 변경
