+++ src/main/kotlin/com/flab/ticketing/performance/repository/PerformanceDateRepository.kt
@@ -1,10 +1,23 @@
 package com.flab.ticketing.performance.repository
 
 import com.flab.ticketing.performance.entity.PerformanceDateTime
-import org.springframework.data.repository.Repository
+import org.springframework.data.jpa.repository.Query
+import org.springframework.data.repository.CrudRepository
+import org.springframework.data.repository.query.Param
+import org.springframework.stereotype.Repository
 
 
-@org.springframework.stereotype.Repository
-interface PerformanceDateRepository : Repository<PerformanceDateTime, Long> {
+@Repository
+interface PerformanceDateRepository : CrudRepository<PerformanceDateTime, Long> {
     fun findByUid(uid: String): PerformanceDateTime?
+
+    @Query(
+        "SELECT pd FROM PerformanceDateTime pd " +
+                "WHERE pd.performance.uid = :performanceUid " +
+                "AND pd.uid = :dateUid"
+    )
+    fun findPerformanceDateTime(
+        @Param("performanceUid") performanceUid: String,
+        @Param("dateUid") dateUid: String
+    ): PerformanceDateTime?
 }
